Allow bugs to be manipulated via regular email messages.
If the very start of the email starts with lines in the form of Key:
Value, then these will be parsed and will allow various bug parameters
to be changed.

This provides a uniform way of manipulating bugs via email messages, and
will allow any manipulations using external tools to be logged in the
form of email messages as well.