Add pre-index and post-index hooks.
These are called when importing a new message. The pre-index hook is
called before the message is indexed, and by returning a non-zero exit
code, it can block the message from being processed. The post-index hook
is called after it has been indexed and all control fields have been
parsed.