Drop the mailbox format and use git-style hash filenames.
Use BLAKE2b to hash the Message-ID of messages, but don't hash the
message itself. The reason is that emails can be modified somewhat when
in transit, but the Message-ID is assumed to be unique and safe from any
mangling. Since the filename is now based on the Message-ID, we don't
have to store the filename explicitly in the database, nor do we need to
create a maildir for each bug.